refactor structure of contestDetails files
