+++ src/Atoms/ContestDetailsLine/ContestDetailsLine.js
