+++ src/Atoms/ContestDetailsLine/ContestDetailsLineStyled.js
