+++ src/Organisms/ContestDetails/ContestDetailsMap/ContestDetailsMap.js
@@ -1,0 +1,11 @@
+import ContestDetailsMapStyled from './ContestDetailsMapStyled';
+
+const ContestDetailsMap = () => {
+  return (
+    <ContestDetailsMapStyled>
+      THERE WILL BE NICE MAP HERE IN THE FUTURE
+    </ContestDetailsMapStyled>
+  );
+};
+
+export default ContestDetailsMap;
