+++ src/Organisms/ContestDetails/ContestDetailsToggler/ContestDetailsTogglerStyled.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
-import COLORS from '../../Styles/varsStyledComponents';
-import FONTS from '../../Styles/fontsStyledComponents';
+import COLORS from '../../../Styles/varsStyledComponents';
+import FONTS from '../../../Styles/fontsStyledComponents';
 
 const ContestDetailsTogglerStyled = styled.div`
   display: flex;
