--- src/Organisms/ContestDetails/Map.js
@@ -1,7 +1,0 @@
-import MapStyled from './MapStyled';
-
-const Map = () => {
-  return <MapStyled>THERE WILL BE NICE MAP HERE IN THE FUTURE</MapStyled>;
-};
-
-export default Map;
